+++ tests/functional/m/missing/missing_yield_doc_Google.rc
@@ -1,0 +1,2 @@
+[MASTER]
+load-plugins = pylint.extensions.docparams
