+++ tests/functional/m/missing/missing_yield_doc_Google.txt
@@ -1,0 +1,1 @@
+redundant-yields-doc:19:0:25:12:my_func:Redundant yields documentation:UNDEFINED
